Simplify css generation and tidy
It was interesting to learn about the various mechanisms that Rake has
for compiling files into targets, but ultimately here we can just get
sass to do the linking and compile a single file ourselves.

The major benefit here is that changes to `code.sass` will trigger the
recompilation of `interblah.css`; previously they would only trigger
recompilation of `code.css`.